count cells are editable; CorrelationManager is the maintainer of all correlation actions; other optimisations
