+++ src/component/panels/SummaryPanel/CorrelationTable/CorrelationTableRow.jsx
@@ -1,8 +1,8 @@
 import lodash from 'lodash';
 import React, { useCallback, useMemo } from 'react';
 
+import { getLabel, getLabels } from '../../../../data/correlation/Utilities';
 import EditableColumn from '../../../elements/EditableColumn';
-import { getLabel, getLabels } from '../Utilities';
 
 const CorrelationTableRow = ({
   additionalColumns,
