+++ src/component/reducer/actions/CorrelationsActions.js
@@ -30,68 +30,49 @@ const handleUnsetTolerance = (state) => {
   });
 };
 
-const handleSetState = (state, correlationsState) => {
-  AnalysisObj.getCorrelationManagerInstance().setState(correlationsState);
+const handleBuildCorrelations = (state, signals1D, signals2D) => {
+  AnalysisObj.getCorrelationManagerInstance().buildValues(signals1D, signals2D);
   return produce(state, (draft) => {
     draft.correlations = AnalysisObj.getCorrelations();
   });
 };
 
-// const handleAddCorrelation = (state, correlation) => {
-//   AnalysisObj.getCorrelationManagerInstance().addValue(correlation);
-//   return produce(state, (draft) => {
-//     draft.correlations = AnalysisObj.getCorrelations();
-//   });
-// };
-
-const handleAddCorrelations = (state, correlations) => {
-  AnalysisObj.getCorrelationManagerInstance().addValues(correlations);
+const handleAddCorrelation = (state, correlation) => {
+  AnalysisObj.getCorrelationManagerInstance().addValue(correlation);
   return produce(state, (draft) => {
     draft.correlations = AnalysisObj.getCorrelations();
   });
 };
 
-// const handleDeleteCorrelation = (state, id) => {
-//   return produce(state, (draft) => {
-//     AnalysisObj.getCorrelationManagerInstance().deleteValue(id);
-//     draft.correlations = AnalysisObj.getCorrelations();
-//   });
-// };
-
-const handleDeleteCorrelations = (state) => {
+const handleDeleteCorrelation = (state, id) => {
+  AnalysisObj.getCorrelationManagerInstance().deleteValue(id);
   return produce(state, (draft) => {
-    AnalysisObj.getCorrelationManagerInstance().deleteValues();
     draft.correlations = AnalysisObj.getCorrelations();
   });
 };
 
-// const handleSetCorrelation = (state, id, correlation) => {
-//   return produce(state, (draft) => {
-//     draft.correlations = AnalysisObj.getCorrelationManagerInstance().setValue(
-//       id,
-//       correlation,
-//     );
-//   });
-// };
+const handleSetCorrelation = (state, id, correlation) => {
+  return produce(state, (draft) => {
+    AnalysisObj.getCorrelationManagerInstance().setValue(id, correlation);
+    draft.correlations = AnalysisObj.getCorrelations();
+  });
+};
 
-// const handleSetCorrelations = (state, correlations) => {
-//   return produce(state, (draft) => {
-//     draft.correlations = AnalysisObj.getCorrelationManagerInstance().setValues(
-//       correlations,
-//     );
-//   });
-// };
+const handleSetCorrelations = (state, correlations) => {
+  return produce(state, (draft) => {
+    AnalysisObj.getCorrelationManagerInstance().setValues(correlations);
+    draft.correlations = AnalysisObj.getCorrelations();
+  });
+};
 
 export {
-  // handleAddCorrelation,
-  handleAddCorrelations,
-  // handleDeleteCorrelation,
-  handleDeleteCorrelations,
-  // handleSetCorrelation,
-  // handleSetCorrelations,
+  handleAddCorrelation,
+  handleBuildCorrelations,
+  handleDeleteCorrelation,
+  handleSetCorrelation,
+  handleSetCorrelations,
   handleSetMF,
   handleUnsetMF,
-  handleSetState,
   handleSetTolerance,
   handleUnsetTolerance,
 };
