+++ src/component/reducer/types/Types.js
@@ -33,17 +33,15 @@ export const SET_ZOOM_FACTOR = 'SET_ZOOM_FACTOR';
 export const ADD_MOLECULE = 'ADD_MOLECULE';
 export const SET_MOLECULE = 'SET_MOLECULE';
 export const DELETE_MOLECULE = 'DELETE_MOLECULE';
-// export const ADD_CORRELATION = 'ADD_CORRELATION';
-export const ADD_CORRELATIONS = 'ADD_CORRELATIONS';
-// export const DELETE_CORRELATION = 'DELETE_CORRELATION';
-export const DELETE_CORRELATIONS = 'DELETE_CORRELATIONS';
-// export const SET_CORRELATION = 'SET_CORRELATION';
-// export const SET_CORRELATIONS = 'SET_CORRELATIONS';
+export const BUILD_CORRELATIONS = 'BUILD_CORRELATIONS';
+export const ADD_CORRELATION = 'ADD_CORRELATION';
+export const DELETE_CORRELATION = 'DELETE_CORRELATION';
+export const SET_CORRELATION = 'SET_CORRELATION';
+export const SET_CORRELATIONS = 'SET_CORRELATIONS';
 export const SET_CORRELATION_MF = 'SET_CORRELATION_MF';
 export const UNSET_CORRELATION_MF = 'UNSET_CORRELATION_MF';
 export const SET_CORRELATION_TOLERANCE = 'SET_CORRELATION_TOLERANCE';
 export const UNSET_CORRELATION_TOLERANCE = 'UNSET_CORRELATION_TOLERANCE';
-export const SET_CORRELATION_STATE = 'SET_CORRELATION_STATE';
 export const DELETE_SPECTRA = 'DELETE_SPECTRA';
 export const CHANGE_SPECTRUM_DISPLAY_VIEW_MODE =
   'CHANGE_SPECTRUM_DISPLAY_VIEW_MODE';
